feat: move maptool things to maptool.py and make it work
